Add basic reservation request functionality
This commit creates a new table for Reservations, which link the User to requested Books (User can have multiple Reservations, but a Book can only have a single Reservation at a time). I've changes the behaviour of the "Reserve" button to not only update the Book's state, but also create a new Reservation object. I've also ensured that a book can be reserved only once.